Make make_validation_schema simpler and useful for extensions
Make it use the merge stratagies properly so that it does not hard code
the field name.

Make the unversioned definitions generate automatically, so are not
hardcoded either.

Try and maintain order so diff is not so bad, even though there are few
reordings that were too much work to avoid.